+++ bucket/covergroup.py
@@ -285,145 +285,3 @@ class Covergroup(CoverBase):
             child_start = child_close.link_across()
         return start.close(self, child=child_close, link=CovRun(point=1), typ=CoverBase)
 
-
-class Covertop(Covergroup):
-    """This is for the top of the coverage tree"""
-
-    def __init__(
-        self, log: logging.Logger | None = None, verbosity: str | int = "INFO"
-    ):
-        if log:
-            assert isinstance(
-                log, logging.Logger
-            ), f"log should be an instance of logging.Logger. Instead got {type(log)}"
-            self.log = log.getChild("bucket")
-            self.log.setLevel(getattr(logging, verbosity, getattr(logging, "INFO")))
-        else:
-            import sys
-
-            self.log = logging.getLogger("bucket")
-            handler = logging.StreamHandler(sys.stdout)
-            logging_verbosity = getattr(logging, verbosity, getattr(logging, "INFO"))
-            handler.setLevel(logging_verbosity)
-            formatter = logging.Formatter(
-                "%(asctime)s - %(name)-20s - %(levelname)s - %(message)s"
-            )
-            handler.setFormatter(formatter)
-            self.log.addHandler(handler)
-        self._init(self.log)
-
-    @validate_call
-    def include_by_function(self, matcher: Callable[[CoverBase], bool]):
-        """
-        Enable coverpoints which match the provided function. Unmatched coverpoints will
-        not have their active state changed, except this is the first filter to be
-        applied, then any coverpoints which do not match are explicitly set to inactive,
-        as the default state is active.
-        Parameters:
-            matcher: A function to match against coverpoint/covergroup data
-        """
-        mismatch_state = None
-        if not self._filter_applied:
-            mismatch_state = False
-        self._apply_filter(matcher, True, mismatch_state)
-        return self
-
-    @validate_call
-    def restrict_by_function(self, matcher: Callable[[CoverBase], bool]):
-        """
-        Filter coverpoints which match the provided function. Those that match will not
-        change their active state, those that don't match will be set to inactive.
-        Parameters:
-            matcher: A function to match against coverpoint/covergroup data
-        """
-        self._apply_filter(matcher, None, False)
-        return self
-
-    @validate_call
-    def exclude_by_function(self, matcher: Callable[[CoverBase], bool]):
-        """
-        Disable coverpoints which match the provided function. Unmatched coverpoints will
-        not have their active state changed.
-        Parameters:
-            matcher: A function to match against coverpoint/covergroup data
-        """
-        self._apply_filter(matcher, False, None)
-        return self
-
-    @validate_call
-    def include_by_name(self, names: MatchStrs):
-        """
-        Enable coverpoints which match the provided names. Unmatched coverpoints will not
-        have their active state changed, except this is the first filter to be applied,
-        then any coverpoints which do not match are explicitly set to inactive, as the
-        default state is active.
-        Parameters:
-            names: A case-insensitive string or string list to match against
-        """
-        return self.include_by_function(self._match_by_name(names))
-
-    @validate_call
-    def restrict_by_name(self, names: MatchStrs):
-        """
-        Filter coverpoints which match the provided names. Those that match will not
-        change their active state, those that don't match will be set to inactive.
-        Parameters:
-            names: A case-insensitive string or string list to match against
-        """
-        return self.exclude_by_function(self._match_by_name(names))
-
-    @validate_call
-    def exclude_by_name(self, names: MatchStrs):
-        """
-        Disable coverpoints which match the provided names. Unmatched coverpoints will
-        not have their active state changed.
-        Parameters:
-            names: A case-insensitive string or string list to match against
-        """
-        return self.exclude_by_function(self._match_by_name(names))
-
-    @validate_call
-    def include_by_tags(self, tags: TagStrs, match_all: bool = False):
-        """
-        Enable coverpoints which match the provided tags. Unmatched coverpoints will not
-        have their active state changed, except this is the first filter to be applied,
-        then any coverpoints which do not match are explicitly set to inactive, as the
-        default state is active.
-        Parameters:
-            tags: Tag(s) to match against
-            match_all: If set, all tags must match.
-                       If cleared, any tags can match (default: False)
-        """
-        return self.include_by_function(self._match_by_tags(tags, match_all))
-
-    @validate_call
-    def restrict_by_tags(self, tags: TagStrs, match_all: bool = False):
-        """
-        Filter coverpoints which match the provided tags. Those that match will not
-        change their active state, those that don't match will be set to inactive.
-        Parameters:
-            tags: Tag(s) to match against
-            match_all: If set, all tags must match. If cleared, any tags can match
-        """
-        return self.exclude_by_function(self._match_by_tags(tags, match_all))
-
-    @validate_call
-    def exclude_by_tags(self, tags: TagStrs, match_all: bool = False):
-        """
-        Disable coverpoints which match the provided tags. Unmatched coverpoints will not
-        have their active state changed.
-        Parameters:
-            tags: Tag(s) to match against
-            match_all: If set, all tags must match. If cleared, any tags can match
-        """
-        return self.exclude_by_function(self._match_by_tags(tags, match_all))
-
-    @validate_call
-    def set_tier_level(self, tier: int):
-        """
-        Filter the coverage tree for coverpoints equal to or less than 'tier'.
-        Parameters:
-            tier: The highest tier level to be set active
-        """
-        self._set_tier_level(tier)
-        return self
